👔  Implement creator and increase leak coverage
