+++ test/unit/checkbox.js
@@ -14,36 +14,22 @@
  * limitations under the License.
  */
 
-describe('MaterialCheckbox', function () {
+var creator = require('./creator');
 
-  function createCheckbox() {
-    var label = document.createElement('label'),
-    input = document.createElement('input'),
-    labelText = document.createElement('span');
-    label.for = 'testCheckbox';
-    label.className = 'mdl-checkbox mdl-js-checkbox';
-    input.type = 'checkbox';
-    input.id = 'testCheckbox';
-    input.className = 'mdl-checkbox__input';
-    label.appendChild(input);
-    labelText.className = 'mdl-checkbox__label';
-    labelText.text = 'Test Checkbox';
-    label.appendChild(labelText);
-    return label;
-  };
+describe('MaterialCheckbox', function () {
 
   it('should be globally available', function () {
     expect(MaterialCheckbox).to.be.a('function');
   });
 
   it('should upgrade successfully', function () {
-    var el = createCheckbox();
+    var el = creator.checkbox();
     componentHandler.upgradeElement(el, 'MaterialCheckbox');
     expect($(el)).to.have.data('upgraded', ',MaterialCheckbox');
   });
 
   it('should get disabled class after being checked', function() {
-    var checkbox = createCheckbox();
+    var checkbox = creator.checkbox();
     componentHandler.upgradeElement(checkbox);
     checkbox.querySelector('input').disabled = true;
     checkbox.MaterialCheckbox.checkDisabled();
@@ -53,7 +39,7 @@ describe('MaterialCheckbox', function () {
   });
 
   it('should get checked class after checking toggle state', function() {
-    var checkbox = createCheckbox();
+    var checkbox = creator.checkbox();
     componentHandler.upgradeElement(checkbox);
     checkbox.querySelector('input').checked = true;
     checkbox.MaterialCheckbox.checkToggleState();
