+++ test/unit/textfield.js
@@ -14,46 +14,28 @@
  * limitations under the License.
  */
 
-describe('MaterialTextfield', function () {
+var creator = require('./creator');
 
-  function createSingleLineTextfield() {
-    var container = document.createElement('div');
-    var input = document.createElement('input');
-    var label = document.createElement('label');
-    var errorMessage = document.createElement('span');
-    container.className = 'mdl-textfield mdl-js-textfield';
-    input.className = 'mdl-textfield__input';
-    input.pattern = '[0-9]';
-    input.id = 'testInput';
-    label.for = input.id;
-    label.className = 'mdl-textfield__label';
-    label.text = 'Number';
-    errorMessage.className = 'mdl-textfield__error';
-    errorMessage.text = 'Positive number only.';
-    container.appendChild(input);
-    container.appendChild(label);
-    container.appendChild(errorMessage);
-    return container;
-  };
+describe('MaterialTextfield', function () {
 
   it('should be globally available', function () {
     expect(MaterialTextfield).to.be.a('function');
   });
 
   it('should upgrade successfully', function () {
-    var el = createSingleLineTextfield();
+    var el = creator.textField(document);
     componentHandler.upgradeElement(el, 'MaterialTextfield');
     expect($(el)).to.have.data('upgraded', ',MaterialTextfield');
   });
 
   it('should be a widget', function () {
-    var el = createSingleLineTextfield();
+    var el = creator.textField(document);
     componentHandler.upgradeElement(el, 'MaterialTextfield');
     expect(el.MaterialTextfield).to.be.a('object');
   });
 
   it('should have public methods available via widget', function () {
-    var el = createSingleLineTextfield();
+    var el = creator.textField(document);
     componentHandler.upgradeElement(el, 'MaterialTextfield');
     var methods = [
       'checkDisabled',
@@ -69,7 +51,7 @@ describe('MaterialTextfield', function () {
   });
 
   it('should be invalid after upgrade if invalid previously', function () {
-    var el = createSingleLineTextfield()
+    var el = creator.textField(document)
     el.classList.add('is-invalid');
     componentHandler.upgradeElement(el);
     expect(el.classList.contains('is-invalid')).to.equal(true);
